move walk type to a separate file
